requirements.txt
tqdm
bcbio-gff
biopython
goatools
twine
#PyQt5 --> for ete3
